Added fps computation+Wall

+++ tests/emptyApp/emptyApp.cpp
@@ -2,7 +2,7 @@
 
 #include <GL/gl.h>
 
-void EmptyApp::render(float elapsedTime)
+void EmptyApp::render(double elapsedTime)
 {
    glClearColor(0, 1, 0, 1);
    glClear(GL_COLOR_BUFFER_BIT);
