+++ tests/emptyApp/emptyApp.h
@@ -9,7 +9,7 @@ public:
   EmptyApp(int resx, int resy, const std::string& name,
            bool fullscreen = false) : ApplicationSDL(resx, resy, name, fullscreen) {}
 
-  void render(float elapsedTime);
+  void render(double elapsedTime);
 };
 
 #endif
